Version 1.1.1
- Refactor the PHP backend.
-- Migrate from MySQL to SQLite using PHP PDO.
-- Add an option in config.php to disable email account verification.

- Upgrade to GWT 2.6.1.

- Fix some typos.

- Update email in about.html.

+++ war/php/activation.php
@@ -7,24 +7,24 @@
 require_once("opendb.php");
 require_once("common.php");
 	
-if (isset($_GET['a'])) {
+if ($use_account_email_activation && isset($_GET['a'])) {
 	$a = $_GET['a'];
 	checkMD5($a);
 	$sql = "SELECT user_id FROM activation WHERE activation_string='$a'";
-	if (! $sth = @mysql_query($sql)) {
-		die(getDBErrorXML());
+	if (! $sth = $dbhandle->query($sql)) {
+		die($db_error_xml);
 	}
-	if ($row = mysql_fetch_row($sth)) {
+	if ($row = $sth->fetch()) {
 		$user_id = $row[0];
 		$sql = "UPDATE user SET active=1 WHERE user_id=$user_id";
-		if (! $sth = @mysql_query($sql)) {
-			die(getDBErrorXML());
+		if (! $sth = $dbhandle->query($sql)) {
+			die($db_error_xml);
 		}
 		echo ('<P><B>Your account is now activated. <B><BR>You can go to the ' .
 			  '<A HREF="http://'.$ldbnhost.'">main page</A>' .
     		  ' and login with your username and password.');
     	$sql = "DELETE FROM activation WHERE activation_string='$a'";
-    	@mysql_query($sql);
+    	$dbhandle->query($sql);
 	} else {
 		echo ("<P>Invalid activation string.</P>");
 	}
